fix: ✂️remove all modded shears
I've noticed that many newbies, when they need to harvest grass or vines, see in JEI that scissors cannot be crafted, and mistakenly think that the modpack developer is forcing them to craft obsidian or other scissors from mods. I hope that if I remove all scissors except for vanilla ones, players will realize that something is wrong here and that they still need to use Kama from Tinkercast Construct.